refactor(oracle): migrate dhcp generator sampler and behavior
