+++ tools/oracle/engine/generator.py
@@ -34,7 +34,6 @@ from .sampling import (
     _IPV4_DOCUMENTATION_NETWORKS,
     _IPV6_DOCUMENTATION_NETWORK,
     _ipv6_next_header_for_stack,
-    _is_ipv4_root_dhcp_stack,
     _json_object,
     _mac_for_domain,
     _next_layer_after,
@@ -100,6 +99,13 @@ from .protocols.ripng import _apply_ripng_behavior
 # backend can encode — the same co-locate-and-re-import pattern as the
 # BGP/OSPF/RIP body injectors above.
 from .protocols.dns import _dns_behavior_emits_raw
+# ``DHCP_OPTION_MATRIX_TOKENS`` moved to the DHCP sampler plugin
+# (``protocols/dhcp.py``) with the rest of the DHCP sampler and ``dhcp_behavior``
+# feature behavior. It is re-imported here because
+# ``tools/oracle/tests/test_dhcp_oracle.py`` imports it from ``generator`` to pin
+# the cross-backend option matrix — the same co-locate-and-re-import pattern as
+# the DNS ``_dns_behavior_emits_raw`` helper above.
+from .protocols.dhcp import DHCP_OPTION_MATRIX_TOKENS
 
 
 SUPPORTED_LAYER_BACKEND = "libcrafter"
@@ -161,17 +167,6 @@ def _ipsec_pinned_crypto() -> JSONObject:
         "integrity_key": {"hex": _IPSEC_PINNED_INTEGRITY_KEY},
     }
 _SUPPORTED_FIELDS: dict[str, set[str]] = {
-    "dhcp": {
-        "op",
-        "hardware_type",
-        "hardware_length",
-        "transaction_id",
-        "flags",
-        "client_ip",
-        "your_ip",
-        "client_hardware_address",
-        "options",
-    },
     "dot11": {
         "addr1",
         "addr2",
@@ -1659,8 +1654,6 @@ class PacketGenerator:
                     grammar=self.grammar,
                 )
                 return
-        if feature == "dhcp_behavior" and "dhcp" in fields:
-            _apply_dhcp_behavior(fields["dhcp"], case=case, behavior=behavior)
 
     def _apply_icmp_live_fields(
         self,
@@ -1796,8 +1789,6 @@ class PacketGenerator:
             return _sample_ah_field(ctx, field_name, domain, field_spec)
         if layer == "ikev2":
             return _sample_ikev2_field(ctx, field_name, domain, field_spec)
-        if layer == "dhcp":
-            return _sample_dhcp_field(ctx, field_name, domain)
         if layer == "radiotap":
             return _sample_radiotap_field(ctx, field_name, domain)
         if layer == "dot11":
@@ -2165,61 +2156,6 @@ def _ikev2_flags_for_domain(domain: object) -> list[str]:
     return ["initiator"]
 
 
-def _sample_dhcp_field(ctx: _SamplingContext, field_name: str, domain: object) -> object:
-    if field_name == "op":
-        return domain
-    if field_name == "hardware_type":
-        return "ethernet"
-    if field_name == "hardware_length":
-        return 6
-    if field_name == "transaction_id":
-        return bounded_int(ctx.rng, 0, (1 << 32) - 1)
-    if field_name == "flags":
-        if domain == "broadcast" and _is_ipv4_root_dhcp_stack(ctx.stack):
-            return "none"
-        return domain
-    if field_name == "client_ip":
-        return "0.0.0.0" if domain == "zero" else ctx.src_ipv4
-    if field_name == "your_ip":
-        return "0.0.0.0" if domain == "zero" else ctx.dst_ipv4
-    if field_name == "client_hardware_address":
-        return ctx.src_mac
-    if field_name == "options":
-        return _dhcp_option_domains(ctx, domain)
-    raise ValueError(f"spec error: unsupported dhcp field sampler: {field_name}")
-
-
-def _dhcp_option_domains(ctx: _SamplingContext, domain: object) -> list[object]:
-    options: list[object] = ["message-type=discover"]
-    boundary = ctx.profile in {"boundary", "fuzz"}
-    if domain == "hostname" and ctx.profile != "smoke":
-        options.append("hostname=libcrafter-oracle")
-    elif domain == "domain_name" and boundary:
-        options.append("domain=example.com")
-    elif domain == "requested_ip" and boundary:
-        options.append(f"requested_addr={ctx.dst_ipv4}")
-    elif domain == "server_id" and boundary:
-        options.append(f"server_id={ctx.src_ipv4}")
-    elif domain == "lease_time" and boundary:
-        options.append(("lease_time", 3600))
-    elif domain == "router" and boundary:
-        options.append(f"router={ctx.src_ipv4}")
-    elif domain == "dns_server" and boundary:
-        options.append(f"name_server={ctx.dst_ipv4}")
-    elif domain == "vendor_class" and boundary:
-        options.append(["vendor_class_id", "6c6962637261667465722d6f7261636c65"])
-    elif domain == "client_identifier" and boundary:
-        options.append(["client_id", "01" + ctx.src_mac.replace(":", "")])
-    elif domain == "classless_static_route" and boundary:
-        options.append(["classless_static_routes", [f"192.0.2.0/24:{ctx.src_ipv4}"]])
-    elif domain == "relay_agent" and boundary:
-        options.append(["relay_agent_information", "0103616263"])
-    elif domain == "parameter_request_list" and boundary:
-        options.append(["param_req_list", [1, 3, 6]])
-    options.append("end")
-    return options
-
-
 # Stable per-code mapping for the representative ICMP destination-unreachable
 # codes named by the live matrix. Each behavior names at most one code so the
 # generated plan exercises a deterministic code byte rather than a random one.
@@ -2394,66 +2330,6 @@ def _is_ospf_smoke_case(case: str) -> bool:
 # ``_apply_ripng_behavior`` body injector is re-imported at the top of this module
 # because the focused rip-smoke profile path calls it directly with ``behavior=""``
 # outside the registry feature loop.
-
-
-# Backend-neutral DHCP option kinds that materialize byte-for-byte through both
-# the Scapy reference backend and the libcrafter adapter, in fixed option order.
-# Kinds the dhcp_behavior option_matrix lists that Scapy cannot encode
-# byte-for-byte (parameter_request_list, vendor_class, client_identifier,
-# classless_static_route, relay_agent, generic) are covered by native libcrafter
-# fixtures rather than this cross-backend matrix; see
-# tools/oracle/specs/layers/dhcp.yaml and features/dhcp-behavior.yaml.
-DHCP_OPTION_MATRIX_TOKENS: tuple[str, ...] = (
-    "message-type=discover",
-    "hostname=libcrafter-oracle",
-    "domain_name=example.com",
-    "requested_ip=192.0.2.100",
-    "server_id=192.0.2.1",
-    "router=192.0.2.1",
-    "domain_name_server=192.0.2.53",
-    "lease_time=3600",
-    "end",
-)
-
-
-def _dhcp_option_matrix() -> list[object]:
-    """Return the deterministic byte-safe DHCP option matrix in option order.
-
-    The matrix spans every cross-backend option kind the dhcp_behavior
-    option_matrix lists as Scapy-byte-safe, in a fixed order, so a single
-    generated plan exercises the whole matrix and the kinds are deterministic
-    for both directions and both DHCP roots.
-    """
-
-    return list(DHCP_OPTION_MATRIX_TOKENS)
-
-
-def _apply_dhcp_behavior(fields: JSONObject, *, case: str, behavior: str) -> None:
-    key = f"{case} {behavior}".replace("_", "-")
-    if "option-matrix" in key:
-        fields["options"] = _dhcp_option_matrix()
-        return
-    if "offer" in key or "ack" in key:
-        fields["op"] = "bootreply"
-        fields["your_ip"] = "192.0.2.100"
-        message = "ack" if "ack" in key else "offer"
-        fields["options"] = [f"message-type={message}", "server_id=192.0.2.1", "end"]
-    elif "nak" in key:
-        fields["op"] = "bootreply"
-        fields["options"] = ["message-type=nak", "server_id=192.0.2.1", "end"]
-    elif "request" in key:
-        fields["options"] = ["message-type=request", "requested_addr=192.0.2.100", "end"]
-    elif "decline" in key:
-        fields["options"] = ["message-type=decline", "requested_addr=192.0.2.100", "end"]
-    elif "inform" in key:
-        fields["options"] = ["message-type=inform", "hostname=libcrafter-oracle", "end"]
-    elif "release" in key:
-        fields["options"] = ["message-type=release", "server_id=192.0.2.1", "end"]
-    elif "force-renew" in key:
-        fields["op"] = "bootreply"
-        fields["options"] = ["message-type=force_renew", "server_id=192.0.2.1", "end"]
-    elif "lease-query" in key:
-        fields["options"] = ["message-type=lease_query", "end"]
 
 
 def _udp_option_feature_tags(case: str, behavior: str | None) -> list[str]:
